requirements.txt
Flask>=2.0

python-dotenv>=0.19

openai>=0.27.0
google-generativeai>=0.2.0
google-api-core>=2.0.0

pandas>=1.3.0
chardet>=4.0.0
plotly>=5.0.0

gunicorn>=20.0

Werkzeug>=2.0
itsdangerous>=2.0
Jinja2>=3.0
click>=8.0

autogen-agentchat
autogen-ext

